week_8/patikagram/requirements.txt
asgiref==3.4.1
Django==4.0
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
